+++ framework/src/native/zcsi/zcsi.rs
@@ -17,7 +17,7 @@ extern "C" {
     pub fn init_secondary(name: *const c_char, nlen: i32, core: i32, vdevs: *mut *const c_char, vdev_count: i32)
         -> i32;
     pub fn init_pmd_port(
-        port: i32,
+        port: u8,
         rxqs: i32,
         txqs: i32,
         rx_cores: *const i32,
@@ -28,18 +28,18 @@ extern "C" {
         tso: i32,
         csumoffload: i32,
     ) -> i32;
-    pub fn free_pmd_port(port: i32) -> i32;
-    pub fn recv_pkts(port: i32, qid: i32, pkts: *mut *mut MBuf, len: i32) -> i32;
-    pub fn send_pkts(port: i32, qid: i32, pkts: *mut *mut MBuf, len: i32) -> i32;
+    pub fn free_pmd_port(port: u8) -> i32;
+    pub fn recv_pkts(port: u8, qid: i32, pkts: *mut *mut MBuf, len: i32) -> i32;
+    pub fn send_pkts(port: u8, qid: i32, pkts: *mut *mut MBuf, len: i32) -> i32;
     pub fn num_pmd_ports() -> i32;
-    pub fn rte_eth_macaddr_get(port: i32, address: *mut MacAddress);
-    pub fn init_bess_eth_ring(ifname: *const c_char, core: i32) -> i32;
-    pub fn init_ovs_eth_ring(iface: i32, core: i32) -> i32;
-    pub fn find_port_with_pci_address(pciaddr: *const c_char) -> i32;
-    pub fn attach_pmd_device(dev: *const c_char) -> i32;
+    pub fn rte_eth_macaddr_get(port: u8, address: *mut MacAddress);
+    pub fn init_bess_eth_ring(ifname: *const c_char, core: i32) -> u8;
+    pub fn init_ovs_eth_ring(iface: i32, core: i32) -> u8;
+    pub fn find_port_with_pci_address(pciaddr: *const c_char) -> u8;
+    pub fn attach_pmd_device(dev: *const c_char) -> u8;
     // FIXME: Generic PMD info
-    pub fn max_rxqs(port: i32) -> i32;
-    pub fn max_txqs(port: i32) -> i32;
+    pub fn max_rxqs(port: u8) -> i32;
+    pub fn max_txqs(port: u8) -> i32;
     pub fn mbuf_alloc() -> *mut MBuf;
     pub fn mbuf_free(buf: *mut MBuf);
     pub fn mbuf_alloc_bulk(array: *mut *mut MBuf, len: u16, cnt: i32) -> i32;
